Threw away NO_VAL conflicts will handle in subsequent commits.

+++ contribs/pam_slurm_adopt/pam_slurm_adopt.c
@@ -358,7 +358,7 @@ static int _rpc_network_callerid(struct callerid_conn *conn, char *user_name,
 		      req.port_src,
 		      user_name);
 		return SLURM_FAILURE;
-	} else if (*job_id == (uint32_t)NO_VAL) {
+	} else if (*job_id == NO_VAL) {
 		debug("From %s port %d as %s: job indeterminate",
 		      ip_src_str,
 		      req.port_src,
