+++ src/sacct/print.c
@@ -1087,7 +1087,7 @@ extern void print_fields(type_t type, void *object)
 					break;
 				}
 			}
-			if (tmp_uint32 == (uint32_t)NO_VAL)
+			if (tmp_uint32 == NO_VAL)
 				tmp_uint32 = NO_VAL;
 			field->print_routine(field,
 					     tmp_uint32,
@@ -1161,7 +1161,7 @@ extern void print_fields(type_t type, void *object)
 					tmp_uint32 = NO_VAL;
 					break;
 				}
-				if (tmp_uint32 == (uint32_t)NO_VAL)
+				if (tmp_uint32 == NO_VAL)
 					tmp_uint32 = NO_VAL;
 			}
 			field->print_routine(field,
@@ -1183,7 +1183,7 @@ extern void print_fields(type_t type, void *object)
 				default:
 					break;
 				}
-				if (tmp_uint64 != (uint64_t)NO_VAL64)
+				if (tmp_uint64 != NO_VAL64)
 					convert_num_unit(
 						(double)tmp_uint64,
 						outbuf, sizeof(outbuf),
@@ -1239,7 +1239,7 @@ extern void print_fields(type_t type, void *object)
 					tmp_uint32 = NO_VAL;
 					break;
 				}
-				if (tmp_uint32 == (uint32_t)NO_VAL)
+				if (tmp_uint32 == NO_VAL)
 					tmp_uint32 = NO_VAL;
 			}
 
@@ -1261,7 +1261,7 @@ extern void print_fields(type_t type, void *object)
 				default:
 					break;
 				}
-				if (tmp_uint64 != (uint64_t)NO_VAL64)
+				if (tmp_uint64 != NO_VAL64)
 					convert_num_unit(
 						(double)tmp_uint64,
 						outbuf, sizeof(outbuf),
@@ -1337,11 +1337,11 @@ extern void print_fields(type_t type, void *object)
 					break;
 				case JOBCOMP:
 				default:
-					tmp_uint64 = (uint64_t)NO_VAL64;
+					tmp_uint64 = NO_VAL64;
 					break;
 				}
 
-				if (tmp_uint64 != (uint64_t)NO_VAL64)
+				if (tmp_uint64 != NO_VAL64)
 					convert_num_unit(
 						(double)tmp_uint64,
 						outbuf, sizeof(outbuf),
@@ -1398,7 +1398,7 @@ extern void print_fields(type_t type, void *object)
 					tmp_uint32 = NO_VAL;
 					break;
 				}
-				if (tmp_uint32 == (uint32_t)NO_VAL)
+				if (tmp_uint32 == NO_VAL)
 					tmp_uint32 = NO_VAL;
 			}
 
@@ -1491,7 +1491,7 @@ extern void print_fields(type_t type, void *object)
 					tmp_uint32 = NO_VAL;
 					break;
 				}
-				if (tmp_uint32 == (uint32_t)NO_VAL)
+				if (tmp_uint32 == NO_VAL)
 					tmp_uint32 = NO_VAL;
 			}
 
@@ -1765,7 +1765,7 @@ extern void print_fields(type_t type, void *object)
 				break;
 			}
 
-			if (tmp_uint64 != (uint64_t)NO_VAL64) {
+			if (tmp_uint64 != NO_VAL64) {
 				bool per_cpu = false;
 				if (tmp_uint64 & MEM_PER_CPU) {
 					tmp_uint64 &= (~MEM_PER_CPU);
@@ -1858,7 +1858,7 @@ extern void print_fields(type_t type, void *object)
 				tmp_uint32 = NO_VAL;
 				break;
 			}
-			if (tmp_uint32 == (uint32_t)NO_VAL)
+			if (tmp_uint32 == NO_VAL)
 				tmp_uint32 = NO_VAL;
 			field->print_routine(field,
 					     tmp_uint32,
